move adapter into mod
