+++ sdk/iot_deviceupdate/examples/get_file.rs
@@ -21,9 +21,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let file_id = env::var("DEVICE_UPDATE_FILE_ID")
         .expect("Missing DEVICE_UPDATE_FILE_ID environment variable.");
 
-    let http_client = azure_core::new_http_client();
     let creds = Arc::new(ClientSecretCredential::new(
-        http_client,
+        azure_core::new_http_client(),
         tenant_id,
         client_id,
         client_secret,
