+++ sdk/iot_deviceupdate/examples/list_providers.rs
@@ -13,9 +13,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let instance_id = env::var("DEVICE_UPDATE_INSTANCE_ID")
         .expect("Missing DEVICE_UPDATE_INSTANCE_ID environment variable.");
 
-    let http_client = azure_core::new_http_client();
     let creds = Arc::new(ClientSecretCredential::new(
-        http_client,
+        azure_core::new_http_client(),
         tenant_id,
         client_id,
         client_secret,
