+++ sdk/iot_deviceupdate/examples/list_versions.rs
@@ -17,9 +17,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let provider = env::var("DEVICE_UPDATE_PROVIDER")
         .expect("Missing DEVICE_UPDATE_PROVIDER environment variable.");
 
-    let http_client = azure_core::new_http_client();
     let creds = Arc::new(ClientSecretCredential::new(
-        http_client,
+        azure_core::new_http_client(),
         tenant_id,
         client_id,
         client_secret,
