+++ sdk/security_keyvault/examples/get_secret_versions.rs
@@ -13,9 +13,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         env::var("KEYVAULT_URL").expect("Missing KEYVAULT_URL environment variable.");
     let secret_name = env::var("SECRET_NAME").expect("Missing SECRET_NAME environment variable.");
 
-    let http_client = azure_core::new_http_client();
     let creds = Arc::new(ClientSecretCredential::new(
-        http_client,
+        azure_core::new_http_client(),
         tenant_id,
         client_id,
         client_secret,
